Get all the moving parts working for gmail
The UX is still pretty raw, but in chrome and ff this extension now succesfully
downloads the raw email for the first message on the page, if any, in gmail,
then dumps the data into a popup. Clicking the submit button sends a POST with
that data to localhost.

Note bundle.js is a concatenation of jquery and contentScript.js because the
built-in fetch api didn't seem to be resolving redirects correctly when called
from the content script. Also, jquery is helpful for finding message ids in the
webmail DOM, and for rendering on the popup page, though we might eventually be
able to drop it as a dependency.