Remove html from messages.json
We escape these strings when injecting them into the dom, so the HTML markup
doesn't work. It's somewhat poor practice to dynamically parse HTML so the
safest thing to do here is just not use markup in translated strings. This also
eases the burden on translators who don't always know how best to deal with html
tags.