Content script refactor / Adding mail providers
Some structural changes to make it more straightforward to add new providers.

Each provider is served by a dedicated function called getRawEmailProviderName.
Each of these functions should return a promise which either resovles to a
string result containing the raw email or rejects with a numeric error status
code.

For now, the popup/frontend determines what the codes mean, though that may have
to change if we add other providers.